// frontend/src/api-generated/models/donHangMuaSamResponse.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * IT Asset Management API
 * Tài liệu API Hệ thống quản lý tài sản CNTT
 * OpenAPI spec version: 1.0
 */
import type { ChiTietDonHangGeneralResponse } from './chiTietDonHangGeneralResponse';

export interface DonHangMuaSamResponse {
  id?: number;
  idDonVi?: number;
  idNhaCungCap?: number;
  tenNhaCungCap?: string;
  /** Tên người lập đơn hàng */
  tenNguoiLap?: string;
  /** Tên người phê duyệt đơn hàng */
  tenNguoiPheDuyet?: string;
  maDonHang?: string;
  soHopDongDinhKem?: string;
  tongTienTruocThue?: number;
  thueVat?: number;
  tongTienSauThue?: number;
  thoiGianGiaoDuKien?: string;
  trangThai?: string;
  ghiChu?: string;
  thoiGianTao?: string;
  thoiGianCapNhat?: string;
  /** Danh sách chi tiết tài sản của đơn hàng mua sắm */
  chiTietTaiSan?: ChiTietDonHangGeneralResponse[];
}