// frontend/src/api-generated/models/donHangMuaSamRequest.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * IT Asset Management API
 * Tài liệu API Hệ thống quản lý tài sản CNTT
 * OpenAPI spec version: 1.0
 */
import type { ChiTietDonHangPhanCungRequest } from './chiTietDonHangPhanCungRequest';
import type { ChiTietDonHangPhanMemRequest } from './chiTietDonHangPhanMemRequest';

export interface DonHangMuaSamRequest {
  idNhaCungCap: number;
  idNguoiLap?: number;
  idNguoiPheDuyet?: number;
  /**
     * @minLength 0
     * @maxLength 50
     */
  maDonHang?: string;
  /**
     * @minLength 0
     * @maxLength 100
     */
  soHopDongDinhKem?: string;
  tongTienTruocThue: number;
  thueVat: number;
  tongTienSauThue: number;
  thoiGianGiaoDuKien?: string;
  ghiChu?: string;
  chiTietPhanCung?: ChiTietDonHangPhanCungRequest[];
  chiTietPhanMem?: ChiTietDonHangPhanMemRequest[];
}